API

TaylorModels.jl

TaylorModels.RTaylorModel1

TaylorModels.TaylorModel1

TaylorModels.TaylorModelN

TaylorModels._rpaar

TaylorModels._rparr

TaylorModels.absorb_remainder

TaylorModels.bound_absrem

TaylorModels.bound_integration

TaylorModels.bound_relrem

TaylorModels.bound_taylor1

TaylorModels.bound_taylor1

TaylorModels.check_existence

TaylorModels.fp_rpa

TaylorModels.iscontractive

TaylorModels.picard_lindelöf

TaylorModels.picard_remainder!

TaylorModels.reducetoorder

TaylorModels.remainder_taylorstep!

TaylorModels.rpa

TaylorModels.rpa

TaylorModels.shrink_for_existance

TaylorModels.tight_remainder

TaylorModels.validated_step!

TaylorSeries.integrate

TaylorModels.RTaylorModel1 — Type.

RTaylorModel1{T,S}

Taylor model in 1 variable, providing a rigurous polynomial approximation (around x_0) and a relative remainder \delta for a function f(x) in one variable, valid in the interval I. Corresponds to definition 2.3.2 of Mioara Joldes' thesis.

Fields:

pol: polynomial approximation, represented as TaylorSeries.Taylor1

rem: the interval bound

x0: expansion point

I: interval over which the Taylor model is defined / valid

The approximation f(x) = \sum_i p_i (x - x_0)^i + \delta (x - x_0)^{n+1} is satisfied for all x\in I; n is the order (degree) of the polynomial p(x).

source

TaylorModels.TaylorModel1 — Type.

TaylorModel1{T,S}

Taylor model in 1 variable, providing a rigurous polynomial approximation (around x_0) and an absolute remainder \Delta for a function f(x) in one variable, valid in the interval I. Corresponds to definition 2.1.3 of Mioara Joldes' thesis.

Fields:

pol: polynomial approximation, represented as TaylorSeries.Taylor1

rem: the interval bound

x0: expansion point

I: interval over which the Taylor model is defined / valid

The approximation f(x) = \sum_{i=0}^n p_i (x - x_0)^i + \Delta is satisfied for all x\in I (0\in\Delta); n is the order (degree) of the polynomial p(x).

source

TaylorModels.TaylorModelN — Type.

TaylorModelN{N,T,S}

Taylor Models with absolute remainder for N independent variables.

source

TaylorModels.bound_absrem — Method.

bound_absrem(f::Function, polf::Taylor1, polfI::Taylor1, x0::Interval, I::Interval)

Bound the absolute remainder of the polynomial approximation of f given by the Taylor polynomial polf around x0 on the interval I. It requires the interval extension polfI of the polynomial that approximates f for the whole interval I, in order to compute the Lagrange remainder.

If polfI[end] has a definite sign, then it is monotonic in the intervals [I.lo, x0] and [x0.hi, I.hi], which is exploited; otherwise, it is used to compute the Lagrange remainder.

source

TaylorModels.bound_relrem — Method.

bound_relrem(f::Function, polf::Taylor1, polfI::Taylor1, x0::Interval, I::Interval)

Bound the relative remainder of the polynomial approximation of f given by the Taylor polynomial polf around x0 on the interval I. It requires an the interval extension polfI of a polynomial that approximates f for the whole interval I, in order to compute the Lagrange remainder.

If polfI[end] has a definite sign, then it is monotonic in the interval I, which is exploited; otherwise, the last coefficients bounds the relative remainder.

source

TaylorModels.fp_rpa — Method.

fp_rpa(tm::TaylorModel1{Interval{T},T})
fp_rpa(tm::RTaylorModel1{Interval{T},T})

Convert a tm TaylorModel1 to a TaylorModel1 whose polynomial coefficients are Float64. The accumulated error is added to the remainder. The mid point of the expansion interval is preferentially rounded down if it is not an exactly representable value.

source

TaylorModels.rpa — Method.

rpa(g::Function, tmf::RTaylorModel1)

Rigurous polynomial approximation (RPA) for the function g using the Taylor Model with absolute remainder tmf. The bound is computed exploiting monotonicity if possible, otherwise, it uses Lagrange bound.

source

TaylorModels.rpa — Method.

rpa(g::Function, tmf::TaylorModel1) rpa(g::Function, tmf::TaylorModelN)

Rigurous polynomial approximation (RPA) for the function g using the Taylor Model with absolute remainder tmf. The bound is computed exploiting monotonicity if possible, otherwise, it uses Lagrange bound.

source

TaylorSeries.integrate — Method.

integrate(a, c0)

Integrates the one-variable Taylor Model (TaylorModel1 or RTaylorModel1) with respect to the independent variable; c0 is the interval representing the integration constant; if omitted it is considered as the zero interval.

source

TaylorModels._rpaar — Method.

_rpaar(f::Function, x0::Interval, I::Interval, _order::Integer)

Rigurous polynomial approximation (RPA) with absolute remainder for the function f on the interval I, using a Taylor expansion around the interval x0 of order _order. The bound is computed by bound_absrem(@ref) exploiting monotonicity if possible, otherwise, it uses Lagrange bound.

source

TaylorModels._rparr — Method.

_rparr(f::Function, x0::Interval, I::Interval, _order::Integer)

Rigurous polynomial approximation (RPA) with relative remainder for the function f on the interval I, using a Taylor expansion around the interval x0 of order _order. The bound is computed by bound_relrem(@ref) exploiting monotonicity if possible, otherwise, it uses the Lagrange coefficient.

source

TaylorModels.absorb_remainder — Method.

absorb_remainder(a::TaylorModelN{N,T,T}) where {N,T}

Returns a TaylorModelN, equivalent to a, such that the remainder is mostly absorbed in the coefficients. The linear shift assumes that a is normalized to the IntervalBox(-1..1, Val(N)).

Ref: Xin Chen, Erika Abraham, and Sriram Sankaranarayanan, "Taylor Model Flowpipe Construction for Non-linear Hybrid Systems", in Real Time Systems Symposium (RTSS), pp. 183-192 (2012), IEEE Press.

source

TaylorModels.bound_integration — Method.

bound_integration(xTM1::TaylorModel1{Interval{S},S}, δt::Interval{S})
bound_integration(xTM1::Vector{TaylorModel1{Interval{S},S}}, δt::Interval{S})

Remainder bound for the integration of a series, given by $δ * remainder(a) + a.pol[order] * δ^(order+1) / (order+1)$. This is tighter that the one used by Berz+Makino, which corresponds to $Δ = aux * remainder(a) + a.pol[order] * aux^(order+1)$.

source

TaylorModels.bound_taylor1 — Method.

bound_taylor1(fT::Taylor1, I::Interval)

Compute a tight polynomial bound for the Taylor polynomial fT in the interval I.

Note: Algorithm 2.1.1 corresponds to evaluate(fT, I) or simply fT(I). This function uses the roots of the derivative offt` to obtain a tighter bound.

source

TaylorModels.bound_taylor1 — Method.

bound_taylor1(fT::Taylor1, fTd::Taylor1, I::Interval)

Compute a tight polynomial bound for the Taylor polynomial fT in the interval I, considering whether its derivative ftd has a definite sign.

source

TaylorModels.check_existence — Method.

check_existence(f, tm::T, xm::T, x0::Interval, x_test::Interval)

Checks that the range of one iterate of the Picard-Lindelöf operator is contained in the a-priori interval x_test (of the dependent variable) that bounds the solution of the ODE defined by f. This function returns an interval of the independent variable where the a-priori solution is warranted to exist; see shrink_for_existance. Here, tm and xm are Taylor Models (TaylorModel1 or RTaylorModel1) defined for the independent and dependent variables, and x0 is the initial condition.

source

TaylorModels.iscontractive — Method.

iscontractive(Δ, Δx)

Checks if Δ .⊂ Δx is satisfied. If `Δ ⊆ Δx is satisfied, it returns true if all cases where == holds corresponds to the zero Interval.

source

TaylorModels.picard_lindelöf — Method.

picard_lindelöf(f, tm::T, xm::T, x0::Interval)
𝒫(f, tm::T, xm::T, x0::Interval)

Returns the application of the Picard-Lindelöf operator associated to the ODE $\dot{x} = f(t,x)$, with initial condition x0. Here, tm and xm are (one-variable) Taylor Models (TaylorModel1 or RTaylorModel1).

𝒫 is an abbreviation of this operator, which is obtained as \mscrP<TAB>.)

source

TaylorModels.picard_remainder! — Method.

picard_remainder!(f!, t, x, dx, xxI, dxxI, δI, δt, Δx, Δ0)

Return the remainder of Picard operator

source

TaylorModels.reducetoorder — Method.

reducetoorder(a::RTaylorModel1, m::Integer)

From a::RTaylorModel1, it returns a RTaylorModel1 of order m.

source

TaylorModels.remainder_taylorstep! — Method.

remainder_taylorstep!(f!, t, x, dx, xI, dxI, δI, δt)

Returns a remainder for the integration step for the dependent variables (x) checking that the solution satisfies the criteria for existence and uniqueness.

source

TaylorModels.shrink_for_existance — Method.

shrink_for_existance(xm::T, t_interval, x_test, max_steps::Integer=20)

Shrinks the a-priori independent-variable interval t_interval so the range of xm, the Taylor Model (TaylorModel1 or RTaylorModel1) associated with the dependent variable, is contained in the a-priori interval x_test. The method used is some sort of bisection. If no independent-variable interval is found within max_steps, an empty interval is returned.

source

TaylorModels.tight_remainder — Method.

tight_remainder(f, tm::T, xm::T, x0::Interval, max_steps::Integer=20)

Returns a Taylor Model for the dependent variable, with a tighter remainder, which is obtained after successive iteration of the Picard-Lindelöf. If the remainder is not tighter (and identity with the former iterate is not obtained) a Taylor Model with an empty interval is returned.

source

TaylorModels.validated_step! — Method.

validated-step!

source